Move k8s migration logic to separate service part
