+++ src/middlewared/middlewared/plugins/kubernetes_linux/migrate.py
@@ -1,0 +1,41 @@
+from middlewared.service import CallError, private, Service
+
+from .utils import applications_ds_name, MIGRATION_NAMING_SCHEMA
+
+
+class KubernetesService(Service):
+
+    @private
+    async def migrate_ix_applications_dataset(self, new_pool, old_pool):
+        snap_details = await self.middleware.call(
+            'zfs.snapshot.create', {
+                'dataset': applications_ds_name(old_pool),
+                'naming_schema': MIGRATION_NAMING_SCHEMA,
+                'recursive': True,
+            }
+        )
+
+        try:
+            old_ds = applications_ds_name(old_pool)
+            new_ds = applications_ds_name(new_pool)
+            migrate_job = await self.middleware.call(
+                'replication.run_onetime', {
+                    'direction': 'PUSH',
+                    'transport': 'LOCAL',
+                    'source_datasets': [old_ds],
+                    'target_dataset': new_ds,
+                    'recursive': True,
+                    'also_include_naming_schema': [MIGRATION_NAMING_SCHEMA],
+                    'retention_policy': 'NONE',
+                    'replicate': True,
+                    'readonly': 'IGNORE',
+                }
+            )
+            await migrate_job.wait()
+            if migrate_job.error:
+                raise CallError(f'Failed to migrate {old_ds} to {new_ds}: {migrate_job.error}')
+        finally:
+            await self.middleware.call('zfs.snapshot.delete', snap_details['id'], {'recursive': True})
+            snap_name = f'{applications_ds_name(new_pool)}@{snap_details["snapshot_name"]}'
+            if await self.middleware.call('zfs.snapshot.query', [['id', '=', snap_name]]):
+                await self.middleware.call('zfs.snapshot.delete', snap_name, {'recursive': True})
